+++ tests/Unit/ZiggyTest.php
@@ -293,6 +293,35 @@ class ZiggyTest extends TestCase
     }
 
     /** @test */
+    public function can_include_subdomain()
+    {
+        app('router')->domain('{team}.ziggy.dev')->post('/', function () {
+            return response()->json(new Ziggy);
+        })->name('home');
+        app('router')->getRoutes()->refreshNameLookups();
+
+        $this->post(route('home', ['team' => 'tgtn']))
+            ->assertJson([
+                'url' => 'http://tgtn.ziggy.dev',
+            ]);
+    }
+
+    /** @test */
+    public function can_include_port()
+    {
+        app('router')->post('/', function () {
+            return response()->json(new Ziggy);
+        })->name('home');
+        app('router')->getRoutes()->refreshNameLookups();
+
+        $this->post('http://ziggy.dev:3000')
+            ->assertJson([
+                'url' => 'http://ziggy.dev:3000',
+                'port' => 3000,
+            ]);
+    }
+
+    /** @test */
     public function can_include_only_middleware_set_in_config()
     {
         config(['ziggy' => [